:sparkles: [open-formulieren/open-forms#1313] Avoid rendering startpage
if doing automatic redirect
